chore(sqlc): add subscription queries and wire engines via build tags
- Add subscription queries to sqlite/postgres sqlc SQL files
- Split subscription impls: raw SQL (!withsqlc) and sqlc (withsqlc)
- Engines default to raw; build with -tags=withsqlc to use sqlc-generated methods
